check-signals
