+++ package.json
@@ -76,6 +76,9 @@
     "@types/qunit": "^2.19.9",
     "autoprefixer": "^10.4.16",
     "backburner.js": "^2.8.0",
+    "express": "^4.18.2",
+    "glint-environment-gxt": "file:./glint-environment-gxt",
+    "happy-dom": "^13.0.6",
     "nyc": "^15.1.0",
     "postcss": "^8.4.33",
     "prettier": "^3.1.1",
@@ -93,16 +96,14 @@
     "vite-plugin-circular-dependency": "^0.2.1",
     "vite-plugin-dts": "^3.7.0",
     "vitest": "^1.1.1",
-    "zx": "^7.2.3",
-    "express": "^4.18.2",
-    "happy-dom": "^13.0.6",
-    "glint-environment-gxt": "file:./glint-environment-gxt"
+    "zx": "^7.2.3"
   },
   "dependencies": {
     "@babel/core": "^7.23.6",
-    "decorator-transforms": "1.1.0",
     "@babel/preset-typescript": "^7.23.3",
     "@glimmer/syntax": "^0.87.1",
-    "content-tag": "^1.2.2"
+    "content-tag": "^1.2.2",
+    "decorator-transforms": "1.1.0",
+    "signal-polyfill": "^0.1.0"
   }
 }
